+++ Orange/classification/base.py
@@ -5,7 +5,7 @@ import scipy
 import bottlechest as bn
 
 import Orange.data
-import Orange.data.preprocess
+import Orange.preprocess.preprocess
 
 __all__ = ["Fitter", "Model", "SklFitter", "SklModel"]
