+++ Orange/classification/logistic_regression.py
@@ -2,7 +2,6 @@ import numpy
 from scipy import sparse
 import sklearn.linear_model as skl_linear_model
 
-import Orange.data.preprocess
 from Orange.classification import SklFitter, SklModel
 
 __all__ = ["LogisticRegressionLearner", "LogisticRegressionClassifier"]
